+++ packages/create-blocks-app/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @aws-blocks/create-blocks-app
 
+## 0.1.17
+
+### Patch Changes
+
+- 50dbd0e: Include the local dev-server script and use public AWS Blocks imports when adding AWS Blocks to an Amplify Gen 2 project.
+
 ## 0.1.16
 
 ### Patch Changes